Convert all C calculation to double instead (no significant impact)

+++ siddon/C_siddon.h
@@ -5,41 +5,41 @@
 typedef struct
 {
   /* steps in radians*/
-  float p1;
-  float p2;
+  double p1;
+  double p2;
   /* size in pixels*/
   int n1;
   int n2;
   /* position of the sun center in pixels (can be a fraction of pixel */
-  float s1;
-  float s2;
+  double s1;
+  double s2;
 }detector;
 
 /* define the position of the detector */
 typedef struct
 {
   /* spherical coordinates*/
-  float lon; /* longitude in radians */
-  float lat; /* latitude in radians */
-  float rol; /* roll angle in radians (angle between the image vertical and the Sun north pole taken positive clockwise )*/
-  float d; /* distance in meters */
+  double lon; /* longitude in radians */
+  double lat; /* latitude in radians */
+  double rol; /* roll angle in radians (angle between the image vertical and the Sun north pole taken positive clockwise )*/
+  double d; /* distance in meters */
   /* cartesian coordinates in carrington */
-  float M[3];
+  double M[3];
 }orbit;
 
 /* define the parameters of the Region of interest of the object */
 typedef struct
 {
   /*size of the object cube in solar radii */
-  float d[3];
+  double d[3];
   /* steps in solar radius : 1 sr = 6.95E8 m */
-  float p[3];
+  double p[3];
   /* size in pixels */
   int n[3];
   /* offset in solar radius : position of the firt pixel to center of rotation */
   /*bounds of the volume in solar radius*/
-  float min[3];
-  float max[3];
+  double min[3];
+  double max[3];
 }RoiO;
 
 void init_C_siddon();
